Assert fp32 for rope embeddings, misc test fixes (#1496)
This wouldn't have caught @savitha-eng's `cast_forward_inputs=True` bug
(that casts these right as they enter the TransformerLayer), but it
turns out our test suite was actually casting these to bfloat16 with
`model.to(bfloat16)` calls 😬 .

This also fixes a few other misc. test failures I saw locally making
sure the esm2 & llama3 recipe and model tests pass.

will require #1495 for tests to pass

---------

+++ bionemo-recipes/recipes/esm2_peft_te/tests/test_infer.py
@@ -33,7 +33,7 @@ def peft_model(recipe_path):
     config.id2label = SS3_ID2LABEL
     config.label2id = SS3_LABEL2ID
 
-    base_model = AutoModelForTokenClassification.from_config(config, trust_remote_code=True)
+    base_model = AutoModelForTokenClassification.from_config(config, trust_remote_code=True, dtype=torch.bfloat16)
 
     lora_config = peft.LoraConfig(
         task_type=peft.TaskType.TOKEN_CLS,
@@ -45,7 +45,7 @@ def peft_model(recipe_path):
     )
 
     model = peft.get_peft_model(base_model, lora_config)
-    model.to(device="cuda", dtype=torch.bfloat16)
+    model.to(device="cuda")
     model.eval()
     return model
